rpca cuda update

+++ rpca_cuda.py
@@ -61,7 +61,11 @@ def shrink(X, tau):
 
 def frobeniusNorm(X):
     Z = gpuarray.empty_like(X)
-    square(X, Z, cuda.In(X.size), block=(10, 10, 1))
+    # N = X.size
+    # N = N.astype(np.int32)
+    # N_gpu = cuda.mem_alloc(.nbytes)
+    # cuda.memcpy_htod(N_gpu, N)
+    square(X, Z, cuda.In(X.size.astype(np.int32)), block=(10, 10, 1))
     accum = gpuarray.sum(Z).get()
     print X.get()
     print Z.get()
